Added functionality to the Flat Maker Button
Also added a function to oscaar that sets the current dirrectory to
.../OSCAAR/
This function assumes that you are operating within the confines of the
directories we have created. If we create an executable this should be
guaranteed. Also, because GUI.py does not share a directory with the
oscaar module. I was forced to change the current directory to code and
call execfile on it because I could not import it. I'm not sure if this
is a problem.